% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_of.R
\name{average_of}
\alias{average_of}
\title{Average of numbers}
\usage{
average_of(x)
}
\arguments{
\item{x}{A vector of input numbers}
}
\value{
a mean value
}
\description{
Average of numbers
}
\examples{
average_of(c(2,3))
}
