% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_ALK.R
\name{CV_ALK}
\alias{CV_ALK}
\title{Function to calculate CV on ALK}
\usage{
CV_ALK(DF)
}
\arguments{
\item{DF}{dataframe with lengths in the firts column and ages on the other columns}
}
\value{
CV by age class and total CV
}
\description{
Function to calculate CV on ALK
}
\examples{
CV_ALK(example_ageLength)
}
