% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_ageLength.R
\docType{data}
\name{example_ageLength}
\alias{example_ageLength}
\title{Example of age-length key}
\format{
An object of class \code{data.frame} with 26 rows and 8 columns.
}
\usage{
example_ageLength
}
\description{
Example of age-length key
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{CA}
\keyword{RDB}
