% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_samples.R
\docType{data}
\name{example_samples}
\alias{example_samples}
\title{Example of CA table (RDB format)}
\format{
An object of class \code{data.frame} with 26989 rows and 39 columns.
}
\usage{
example_samples
}
\description{
Example of CA table (RDB format)
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{CA}
\keyword{RDB}
