% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_numeric.R
\name{make_summary_numeric}
\alias{make_summary_numeric}
\title{Summary statistics calculation: mean length, se, median, min, max,  n classes sampled}
\usage{
make_summary_numeric(df1, variable, a, b)
}
\arguments{
\item{df1}{data frame of sampling data}

\item{variable}{"lenCls"}

\item{a}{coefficient of length-weight relationship}

\item{b}{coefficient of length-weight relationship}
}
\value{
table reporting the different estimates
}
\description{
Summary statistics calculation: mean length, se, median, min, max,  n classes sampled
}
\examples{
make_summary_numeric(example_samples,"lenCls",a=0.0006,b=3)

}
