% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWCV.R
\name{MWCV}
\alias{MWCV}
\title{Summary statistics calculation: mean length, se, median, min, max, MWCV, n classes sampled}
\usage{
MWCV(df1, variable)
}
\arguments{
\item{variable}{"lenCls"}

\item{y}{data frame of sampling data}
}
\value{
MWCV
}
\description{
Summary statistics calculation: mean length, se, median, min, max, MWCV, n classes sampled
}
\examples{
MWCV(example_samples,"lenCls")


}
