% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\name{setData}
\alias{setData}
\title{Set data for ADV analysis}
\usage{
setData(
  data.hl,
  data.sl,
  data.hh,
  selected.species = NULL,
  selected.country = NULL,
  selected.year = NULL,
  selected.quarter = NULL,
  selected.area = NULL,
  selected.metier = NULL,
  selected.sampType = NULL,
  level = "trip",
  delta = 1
)
}
\arguments{
\item{data.hl}{dataframe. HL data}

\item{data.sl}{dataframe. SL data}

\item{data.hh}{dataframe. HH data}

\item{selected.species}{String. The species}

\item{selected.country}{String. The country code}

\item{selected.year}{Numeric. The year}

\item{selected.quarter}{Numeric. The quarter}

\item{selected.area}{String. The year}

\item{selected.metier}{String. The metier (optional)}

\item{selected.sampType}{String. The sample type (optional)}

\item{level}{String. Aggregation level. Default is "trip"}

\item{delta}{Numeric. The bin width (?). Default is 1,}
}
\value{
A data.frame
}
\description{
Sets the data for use in ADV analysis....
}
\examples{

}
