% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data - short title}
\format{
Describe the format or structure of the data and what it is. E.g. a
data frame with 2 columns of some random numbers.
}
\source{
Who added it or where it came from. E.g. Data comes from
\href{http://www.ices.dk}{ICES} (this is how you add a weblink)
}
\usage{
data(test_data) or my_function(test_data)
}
\description{
What the dataset is
}
\details{
Optional. More details about the data. E.g. what each column is,
what you might use it for, which functions it goes with, etc. Delete
section if not using.
}
\examples{
# Example of how to use the data in a function (optional, but useful)
my_function(test_data)
}
\seealso{
\code{\link{my_function}} - Optional link to the function or functions it goes
with. Delete section if not using.
}
\keyword{datasets}
