% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maturity_ogive.R
\name{Maturity_ogive}
\alias{Maturity_ogive}
\title{Maturity ogive parameters estimation (L25, L50, L75)}
\usage{
Maturity_ogive(data = simulation_biodata)
}
\arguments{
\item{data}{text}
}
\value{
L25 (...) , L50, L75
}
\description{
Maturity ogive parameters estimation (L25, L50, L75)
}
\examples{


}
